// laba11/src/features/posts/types.ts
export interface Post {
    userId: number;
    id: number;
    title: string;
    body: string;
  }
  
  export interface NewPost {
    userId?: number; 
    title: string;
    body: string;
  }

  export interface PostsState {
    posts: Post[];
    status: 'idle' | 'loading' | 'succeeded' | 'failed';
    error: string | null;
  }